Move root/helpers into app
