+++ src/Index.tsx
@@ -12,9 +12,9 @@ import { initi18n } from './app/i18n';
 import { polyfill } from 'mobile-drag-drop';
 import 'mobile-drag-drop/default.css';
 
-import registerServiceWorker from './register-service-worker';
-import { safariTouchFix } from './safari-touch-fix';
-import Root from './Root';
+import registerServiceWorker from './app/register-service-worker';
+import { safariTouchFix } from './app/safari-touch-fix';
+import Root from './app/Root';
 import updateCSSVariables from './app/css-variables';
 import setupRateLimiter from './app/bungie-api/rate-limit-config';
 import { SyncService } from './app/storage/sync.service';
