+++ src/app/i18next-t.ts
@@ -1,5 +1,9 @@
 import i18next from 'i18next';
 
+/**
+ * Wrap the t function so we can import a properly typed version. The default library won't let you.
+ * This also includes a performance optimization that skips interpolation when no options are provided.
+ */
 export const t = (
   key: string | string[],
   options?: string | i18next.TOptions<i18next.StringMap> | undefined
