+++ src/app/item-picker/ItemPickerContainer.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Subscriptions } from '../rx-utils';
-import { router } from '../../router';
+import { router } from '../router';
 import { ItemPickerState, showItemPicker$ } from './item-picker';
 import ItemPicker from './ItemPicker';
