+++ src/app/redux-utils.ts
@@ -1,6 +1,11 @@
 import { RootState } from './store/reducers';
 import store from './store/store';
 
+/**
+ * Add an observer to a Redux store. Whenever the data selected by the selector parameter
+ * changes, the onChange function will be called. This allows us to react to changes in the
+ * store and update other things, such as saving to IndexedDB or updating CSS variables.
+ */
 export function observeStore<T>(
   select: (state: RootState) => T,
   onChange: (currentState: T, newState: T) => void
