+++ src/app/safari-touch-fix.ts
