+++ src/app/shell/Link.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { DestinyAccount } from '../accounts/destiny-account.service';
 import { UISrefActive, UISref } from '@uirouter/react';
-import { router } from '../../router';
+import { router } from '../router';
 
 interface Props {
   account?: DestinyAccount;
